Added in Functionality for movement in relative to magnetism.
Movement should be intuitive on the different faces of a magnetic platform now. Keep in mind that Right now I'm calling magnetic platforms "platform" but we may very well add an actual platform layer and just need to rename layer 3. Going to have to figure this out more but for now it's a good starting point.